implement JWT authentication functionality

+++ frontend/src/components/Blogs/Blogs.jsx
@@ -2,6 +2,9 @@
 import React, {useEffect, useState} from 'react'
 import { Link } from 'react-router-dom';
 
+/** React bootstrap components */
+import { Col } from 'react-bootstrap';
+
 /** Components */
 import Header from "../Layout/Header";
 
@@ -30,47 +33,51 @@ const Blogs = ()=> {
  	}
 
 	const deleteHandler= async (bid) => {
-        const request = {
-            method: 'DELETE',
-            headers: {
-				'Content-Type':	'application/json',
-				'Accept':		'application/json'
+		if( window.confirm("Are you sure?") )
+		{
+			const request = {
+			    method: 'DELETE',
+			    headers: {
+					'Content-Type':	'application/json',
+					'Accept':		'application/json',
+					'Authorization': localStorage.getItem('user_jwt'),
+				}
 			}
-        }
-        const url = process.env.REACT_APP_API_URL + `blog/${bid}`;
-        const response = await fetch(url, request);
-		const result = await response.json();
-		console.log(result);
-		getBlogs()
+			const url = process.env.REACT_APP_API_URL + `blog/${bid}`;
+			const response = await fetch(url, request);
+			await response.json();
+			getBlogs()
+		}
 	}
 
 	const listGenerator = (value, index, array) => {
-
 		const date = new Date(value.createdAt);
 		const finalDate = date.getDate() + '-' + (date.getMonth()+1) +'-' + date.getFullYear();
 		return(
-			<div key={index} className='col-md-4'>
-				<div className="card mb-4 box-shadow">
-					<img className="card-img-top" src={process.env.REACT_APP_API_URL + value.thumbnail} alt="blog-view" />
-						<div className="card-body">
-							<h4>{value.title}</h4>
-							<p className="card-text">{value.content}</p>
-							<div className="d-flex justify-content-between align-items-center">
-								{
-									localStorage.getItem("user_jwt") ? (
-										<div className="btn-group">
-											<Link className='btn btn-sm btn-outline-secondary' to={`/blog/${value._id}`}>Edit</Link>
-											<Link className='btn btn-sm btn-outline-secondary' to='#!' onClick={ () => deleteHandler(value._id)}>Delete</Link>
-										</div>
-									) : ''
-								}
-							<small className="text-muted">
-								{finalDate}
-							</small>
+			<>
+				<Col md={4} key={index}>
+					<div className="card mb-4 box-shadow">
+						<img className="card-img-top" src={process.env.REACT_APP_API_URL + value.thumbnail} alt="blog-view" />
+							<div className="card-body">
+								<h4>{value.title}</h4>
+								<p className="card-text">{value.content}</p>
+								<div className="d-flex justify-content-between align-items-center">
+									{
+										localStorage.getItem("user_jwt") ? (
+											<div className="btn-group">
+												<Link className='btn btn-sm btn-outline-secondary' to={`/blog/${value._id}`}>Edit</Link>
+												<Link className='btn btn-sm btn-outline-secondary' to='#!' onClick={ () => deleteHandler(value._id)}>Delete</Link>
+											</div>
+										) : ''
+									}
+								<small className="text-muted">
+									{finalDate}
+								</small>
+							</div>
 						</div>
 					</div>
-				</div>
-			</div>
+				</Col>
+			</>
 		);
 	}
